Make trait for absolutely identical code. Add key "ignoreTables" for databases with mysql driver.
